Upgrades: check for empty start & end.
This commit ensures that when migrating from 1.x to 2.x, that no Events go missing from the admin area UI. We do this by checking for and avoiding empty start & end values, and changing them to the most valid value (either "now" or make the end equal the start).

Fixes #33.