Issue/33 - Prevent empty start & end values (#34)
* Bail early if no array to get a post ID from.

This commit avoids a debug notice in Sugar Calendar when another plugin mangles the map_meta_cap filter by no longer returning an array.

* Upgrades: check for empty start & end.

This commit ensures that when migrating from 1.x to 2.x, that no Events go missing from the admin area UI. We do this by checking for and avoiding empty start & end values, and changing them to the most valid value (either "now" or make the end equal the start).

Fixes #33.

* Upgrades: Use metabox functions to avoid complex code duplication.

Also, only loop through posts for 2.x migration, not for flushing rewrite rules. This fixes a bug when both upgrades are being done back-to-back.

Fixes #33.