Build Tools: add Grunt tasks for building Lite & Standard versions.
This commit is a first pass at automating the work necessary to build the 2 different versions of Sugar Calendar. It includes grunt-git and registers a few helper tasks to wrap around some file system actions.

Also runs "grunt update" to make sure all versions and translations are up to date for this branch.