Conditionally allow time zone edits on Events even when setting is off.
This commit checks for existing values, and allows them to be changed if they are already set. This is useful for situations where the feature was enabled previously, but is disabled now, and the data still exists for the Event. Hiding the interface in that situation would not allow it to be changed, which means the feature would need to be toggled on and off to do so.