Switch from date_i18n() to gmdate() in Event::format_date().
OK, this might be a weird one... The problem with date_i18n() (and now wp_date()) is that it always uses and assumes wp_timezone(), which in no longer guaranteed. In addition, those functions do not accept a timezone parameter, and they will also result in localizing the output, which may not actually be desirable.

In my testing, the actual goal with this method is simply to format the date output, not to translate it.

And while it's convenient that WordPress sets the PHP to UTC, that also is not guaranteed (thanks to plugins being able to override this anytime) so date() is unlikely to be desirable here either.

This method (and start_date() and end_date(), too) are not used everywhere they could (or should) be, making this an easy change with no real consequences, and possibly some bug avoidance for anyone who is using them and potentially ends up with unexpectedly translated output.