Holy schnikes I think it's working.
Introduce sugar_calendar_get_timezone_diff_multi(), and use it in sugar_calendar_get_timezone_diff().

This new function is the backbone of all multi-time-zone calculations going forward, and is used to calculate the time differential of multiple zones. This is useful and necessary when calendars, events, and users all have their own "home zones" and need to be adjusted when being rendered inside a GUI.

Also adds some first pass PHP Unit tests. More to follow just to make sure this is all really correct.