Improve timezone sanitization and UI in meta-box:
This commit makes sure that time zone fields are visible (or not) based on not just the Setting, but also the Event being edited.

If multiple time zones exist for an Event, those need to be editable, regardless of the setting being "Off" or "Single". Same with single zones if set to "Off", and with single zones when being set to "Multi".

If the start & end times match, this changes the behavior of how time zones work and save. More to do on this later, but for now this check is also added.