Deprecate overlaps() and introduce intersects().
This commit is necessary to ensure proper time zone support for recurring Event sequences.

The new Event::intersects() method uses DateTime objects (with DateTimeZone objects) instead of gmdate/gmmktime and manually mathing the offsets, which I was failing at making work if I'm being totally honest here.

What happens now, is that every "cell" or individual "boundary" carries with it the DateTimeZone object for the current time zone view preference, and the new intersects() method is able to directly compare that DateTime object to the ones created locally using the Start & End Time Zones for the Event itself. Certain restrictions apply to this rule, though, specifically for All Day Events and Floating Time Zone Views, where it's important that the Event Time Zones are ignored if they are not empty.

This allows for simplifying the "is_item_for_cell()" method, as that logic now lives in "intersects()".

Standard Recurring (and Advanced Recurring) will be updated to use this new method and filter.

A future iteration of this could be even more clean, but this has already taken longer than I'd planned, and it can be revisited when necessary.

Thank you for coming to my Ted Talk.