Shortcodes: add support for all start-of-week values.
This commit introduces two helper functions to allow for correctly calculating the offset of shortcode output for all supported views (excluding "day" and "4day").

It also adds support for a "sow" shortcode attribute, allowing for start-of-week to be overridden on a per-shortcode basis, including AJAX navigation. This requires adding a new parameter to mulitple functions, so that it may be passed and/or overridden.

While poking around the shortcode attributes, I'm also adding support for "month" and "year" overrides, which oddly has never been supported but was easily added in. Shh. Maybe no one will notice! 🤫