Berlin\Queries\Date: Add $now support, and 'second' support.
This commit enhances `build_mysql_datetime()` by allowing a bit more granularity with what the "current" time is perceived to be, as well as adding direct support for 'Y-m-d H:i:s' formatted string values. It also adds direct support for a unix timestamp as the $datetime value, treating it similarly to the numeric string.

This change facilitates some changes to List Tables, and will be merged upstream to Berlin imminently.