List Tables: final step in pre-fetching Status counts.
These changes add an $item_counts variable to the Base list table class, that is used to store an array of pre-fetched counts, based on the query->items that intersect with the current view boundaries.

It introduces a new 'all_items' cell property that holds onto non-filtered items, or items that are not specific to the current status. (In retrospect, "filtered_items" may not have been the best variable name.)

It also moves some code to override the List Table, out of the "List mode" class and into the Base class, allowing arguments to override the list table just-in-time. This is no longer used internally, but may be used in a future version by add-ons.

Lastly, it also includes some whitespace fixes to week/day mode list table classes.