Recur: support Events with "no known end". (#215)
This change ensures that the Recur class does not error when an Event duration time is unknown, or has no defined end time. In this instance, when the start equals the end, a 0 value duration_time is set. Previous logic would discard this Event as erroneous, which was itself erroneous.

See #214.